requirements.txt
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.21.0
Pillow>=9.0.0 